refactor(frontend): modified the component names to be section-agnostic, so thy can be reused.
