+++ src/app/(frontend)/components/AboutUsCard.tsx
