+++ src/app/(frontend)/page.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import './styles.css'
-import PartnersGrid from './components/PartnersGrid'
+import Grid from './components/Grid'
 
 export default function HomePage() {
   return (
     <div className="min-h-screen bg-white text-black">
-      <PartnersGrid />
+      <Grid />
     </div>
   )
 }
